WIP of Elixir API tester
It has been a few months since I looked at this - looking at http_api_tester I'm a bit confused as to what I was programming towards. Conceptually the goal was to make a Elixir client that could perform heavy concurrent HTTP testing. I will either come back to this or implement it in a better place - for now I want to at least get it out of my local.